refactor: minimize pp program dependencies (#521)
