+++ Dockerfile
@@ -4,9 +4,10 @@ ARG CIRCUIT_ARTIFACTS_URL_BASE=https://sp1-circuits.s3-us-east-2.amazonaws.com
 ARG CIRCUIT_TYPE=plonk
 ARG CIRCUIT_VERSION=v3.0.0
 ARG PROTOC_VERSION=28.2
+ARG CHEF_VERSION=0.1.68
 
 USER root
-RUN cargo install cargo-chef
+RUN cargo install --version ${CHEF_VERSION} cargo-chef
 WORKDIR /app
 
 FROM chef AS planner
