+++ crates/agglayer-aggregator-notifier/src/packer/mod.rs
@@ -16,6 +16,7 @@ use ethers::{
     providers::PendingTransaction,
     types::{TransactionReceipt, H256, U256, U64},
 };
+use pessimistic_proof::proof::DisplayToHex;
 use pessimistic_proof::PessimisticProofOutput;
 use tracing::{debug, error, info, instrument, warn};
 
@@ -162,7 +163,7 @@ where
         let contract_call = self
             .l1_rpc
             .build_verify_pessimistic_trusted_aggregator_call(
-                *output.origin_network,
+                output.origin_network,
                 l_1_info_tree_leaf_count,
                 *output.new_local_exit_root,
                 *output.new_pessimistic_root,
