+++ crates/agglayer-node/Cargo.toml
@@ -2,6 +2,7 @@
 name = "agglayer-node"
 version.workspace = true
 edition.workspace = true
+license.workspace = true
 
 [lints]
 workspace = true
