+++ crates/pessimistic-proof-core/Cargo.toml
@@ -1,27 +1,29 @@
 [package]
-name = "pessimistic-proof"
-version.workspace = true
+name = "pessimistic-proof-core"
+version = "0.1.0"
 edition.workspace = true
 license.workspace = true
 
 [dependencies]
 agglayer-primitives.workspace = true
+
 bincode.workspace = true
-hex-literal = "0.4"
 hex.workspace = true
-pessimistic-proof-core.workspace = true
+hex-literal = "0.4"
+tracing.workspace = true
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["arbitrary_precision"] }
 serde_with = { version = "3" }
 thiserror.workspace = true
-
-[dev-dependencies]
-pessimistic-proof-core = { workspace = true, features = ["testutils"] }
-rand = "0.8.5"
-rs_merkle = { version = "1.4", default-features = false }
 tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", branch = "patch-v2.0.2", features = [
     "keccak",
 ] }
+rand = { version = "0.8.5", optional = true }
 
-[lints]
-workspace = true
+[dev-dependencies]
+rs_merkle = { version = "1.4", default-features = false }
+rand = "0.8.5"
+
+[features]
+default = []
+testutils = ["rand"]
