+++ crates/pessimistic-proof-program/Cargo.toml
@@ -6,7 +6,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 bincode = "1.3.3"
-pessimistic-proof = { path = "../pessimistic-proof" }
+pessimistic-proof-core = { path = "../pessimistic-proof-core", default-features = false }
 sp1-zkvm = "1.1.1"
 
 [build-dependencies]
