+++ crates/pessimistic-proof-program/src/main.rs
@@ -1,13 +1,13 @@
 #![no_main]
 
 use bincode::Options;
-use pessimistic_proof::local_exit_tree::hasher::Keccak256Hasher;
-use pessimistic_proof::multi_batch_header::MultiBatchHeader;
-use pessimistic_proof::{generate_pessimistic_proof, LocalNetworkState, PessimisticProofOutput};
+use pessimistic_proof_core::local_exit_tree::hasher::Keccak256Hasher;
+use pessimistic_proof_core::multi_batch_header::MultiBatchHeader;
+use pessimistic_proof_core::{generate_pessimistic_proof, NetworkState, PessimisticProofOutput};
 
 sp1_zkvm::entrypoint!(main);
 pub fn main() {
-    let initial_state = sp1_zkvm::io::read::<LocalNetworkState>();
+    let initial_state = sp1_zkvm::io::read::<NetworkState>();
     let batch_header = sp1_zkvm::io::read::<MultiBatchHeader<Keccak256Hasher>>();
 
     let outputs = generate_pessimistic_proof(initial_state, &batch_header).unwrap();
