+++ crates/pessimistic-proof-test-suite/src/event_data.rs
@@ -2,7 +2,8 @@ use std::{fs::File, io::BufReader};
 
 use agglayer_types::primitives::U256;
 use base64::{engine::general_purpose::STANDARD, Engine};
-use pessimistic_proof::bridge_exit::{BridgeExit, TokenInfo};
+use pessimistic_proof::bridge_exit::BridgeExit;
+use pessimistic_proof::bridge_exit::TokenInfo;
 use serde::{de::DeserializeOwned, Deserialize, Deserializer};
 use serde_json::Number;
 
@@ -72,7 +73,7 @@ impl From<DepositEventData> for BridgeExit {
         Self {
             leaf_type: deposit_event_data.leaf_type.try_into().unwrap(),
             token_info: TokenInfo {
-                origin_network: deposit_event_data.origin_network.into(),
+                origin_network: deposit_event_data.origin_network,
                 origin_token_address: deposit_event_data.origin_address.parse().unwrap(),
             },
             dest_network: deposit_event_data.destination_network.into(),
