+++ crates/pessimistic-proof-test-suite/tests/e2e_local_pp.rs
@@ -1,5 +1,7 @@
 use agglayer_types::primitives::U256;
-use pessimistic_proof::{bridge_exit::TokenInfo, generate_pessimistic_proof, LocalNetworkState};
+use pessimistic_proof::bridge_exit::TokenInfo;
+use pessimistic_proof::core::generate_pessimistic_proof;
+use pessimistic_proof::local_state::LocalNetworkState;
 use pessimistic_proof_test_suite::{
     forest::Forest,
     sample_data::{ETH, USDC},
