+++ crates/pessimistic-proof/src/bridge_exit.rs
@@ -1,58 +1,21 @@
-#![allow(clippy::too_many_arguments)]
-
 use std::{fmt::Display, ops::Deref};
 
-use agglayer_primitives::{address, Address, U256};
-use hex_literal::hex;
+use agglayer_primitives::{Address, U256};
+pub use pessimistic_proof_core::bridge_exit::{LeafType, TokenInfo};
+use pessimistic_proof_core::{
+    bridge_exit::L1_ETH,
+    keccak::{digest::Digest, keccak256, keccak256_combine},
+};
 use serde::{Deserialize, Serialize};
 
-use crate::keccak::{digest::Digest, keccak256, keccak256_combine};
+use crate::utils::Hashable;
 
-pub(crate) const L1_NETWORK_ID: NetworkId = NetworkId(0);
-pub(crate) const L1_ETH: TokenInfo = TokenInfo {
-    origin_network: L1_NETWORK_ID,
-    origin_token_address: address!("0000000000000000000000000000000000000000"),
-};
-
-/// Encapsulates the information to uniquely identify a token on the origin
-/// network.
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize, Copy)]
-pub struct TokenInfo {
-    /// Network which the token originates from
-    pub origin_network: NetworkId,
-    /// The address of the token on the origin network
-    pub origin_token_address: Address,
-}
-
-impl TokenInfo {
-    /// Computes the Keccak digest of [`TokenInfo`].
-    pub fn hash(&self) -> Digest {
+impl Hashable for TokenInfo {
+    fn hash(&self) -> Digest {
         keccak256_combine([
             &self.origin_network.to_be_bytes(),
             self.origin_token_address.as_slice(),
         ])
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub enum LeafType {
-    Transfer = 0,
-    Message = 1,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, thiserror::Error)]
-#[error("Invalid leaf type number")]
-pub struct LeafTypeFromU8Error;
-
-impl TryFrom<u8> for LeafType {
-    type Error = LeafTypeFromU8Error;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0 => Ok(Self::Transfer),
-            1 => Ok(Self::Message),
-            _ => Err(LeafTypeFromU8Error),
-        }
     }
 }
 
@@ -76,9 +39,18 @@ pub struct BridgeExit {
     pub metadata: Option<Digest>,
 }
 
-const EMPTY_METADATA_HASH: Digest = Digest(hex!(
-    "c5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470"
-));
+impl From<BridgeExit> for pessimistic_proof_core::bridge_exit::BridgeExit {
+    fn from(value: BridgeExit) -> Self {
+        Self {
+            leaf_type: value.leaf_type,
+            token_info: value.token_info,
+            dest_network: value.dest_network.into(),
+            dest_address: value.dest_address,
+            amount: value.amount,
+            metadata: value.metadata,
+        }
+    }
+}
 
 impl BridgeExit {
     /// Creates a new [`BridgeExit`].
@@ -94,7 +66,7 @@ impl BridgeExit {
         Self {
             leaf_type,
             token_info: TokenInfo {
-                origin_network,
+                origin_network: *origin_network,
                 origin_token_address,
             },
             dest_network,
@@ -104,23 +76,26 @@ impl BridgeExit {
         }
     }
 
-    /// Hashes the [`BridgeExit`] to be inserted in a
-    /// [`crate::local_exit_tree::LocalExitTree`].
-    pub fn hash(&self) -> Digest {
-        keccak256_combine([
-            [self.leaf_type as u8].as_slice(),
-            &u32::to_be_bytes(self.token_info.origin_network.into()),
-            self.token_info.origin_token_address.as_slice(),
-            &u32::to_be_bytes(self.dest_network.into()),
-            self.dest_address.as_slice(),
-            &self.amount.to_be_bytes::<32>(),
-            &self.metadata.unwrap_or(EMPTY_METADATA_HASH).0,
-        ])
-    }
-
     pub fn is_transfer(&self) -> bool {
         self.leaf_type == LeafType::Transfer
     }
+}
+
+impl Hashable for BridgeExit {
+    fn hash(&self) -> Digest {
+        pessimistic_proof_core::bridge_exit::bridge_exit_hasher(
+            self.leaf_type as u8,
+            self.token_info.origin_network,
+            self.token_info.origin_token_address,
+            *self.dest_network,
+            self.dest_address,
+            self.amount,
+            self.metadata,
+        )
+    }
+}
+
+impl BridgeExit {
     pub fn is_message(&self) -> bool {
         self.leaf_type == LeafType::Message
     }
@@ -148,7 +123,7 @@ impl Display for NetworkId {
 
 impl NetworkId {
     pub const BITS: usize = u32::BITS as usize;
-    pub fn new(value: u32) -> Self {
+    pub const fn new(value: u32) -> Self {
         Self(value)
     }
 }
@@ -175,8 +150,10 @@ impl Deref for NetworkId {
 
 #[cfg(test)]
 mod tests {
+    use pessimistic_proof_core::local_exit_tree::hasher::Keccak256Hasher;
+
     use super::*;
-    use crate::local_exit_tree::{hasher::Keccak256Hasher, LocalExitTree};
+    use crate::local_exit_tree::LocalExitTree;
 
     #[test]
     fn test_deposit_hash() {
