+++ crates/pessimistic-proof/src/lib.rs
@@ -1,9 +1,7 @@
-pub mod keccak;
 pub mod local_exit_tree;
 
-mod proof;
-pub use agglayer_primitives::{Address, Signature, U256};
-pub use proof::{generate_pessimistic_proof, PessimisticProofOutput, ProofError};
+pub mod proof;
+pub use proof::PessimisticProofOutput;
 
 pub mod local_balance_tree;
 
@@ -12,8 +10,19 @@ pub mod bridge_exit;
 pub mod global_index;
 pub mod imported_bridge_exit;
 pub mod local_state;
-pub mod multi_batch_header;
 pub mod nullifier_tree;
 pub mod utils;
 
 pub use local_state::LocalNetworkState;
+
+pub mod keccak {
+    pub use pessimistic_proof_core::keccak::*;
+}
+
+pub use pessimistic_proof_core::local_state::NetworkState;
+pub use pessimistic_proof_core::multi_batch_header;
+pub use pessimistic_proof_core::proof::ProofError;
+
+pub mod core {
+    pub use pessimistic_proof_core::generate_pessimistic_proof;
+}
