+++ crates/pessimistic-proof/src/local_balance_tree.rs
@@ -1,19 +1,12 @@
 use std::hash::Hash;
 
-use agglayer_primitives::U256;
+pub use pessimistic_proof_core::local_balance_tree::{LocalBalancePath, LOCAL_BALANCE_TREE_DEPTH};
+use pessimistic_proof_core::{
+    local_exit_tree::hasher::{Hasher, Keccak256Hasher},
+    utils::FromU256,
+};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use serde_with::serde_as;
-
-use crate::{
-    bridge_exit::TokenInfo,
-    local_exit_tree::hasher::Hasher,
-    utils::smt::{SmtMerkleProof, ToBits},
-    ProofError,
-};
-
-/// The key is [`TokenInfo`] which can be packed into 192 bits (32 for network
-/// id and 160 for token address).
-pub const LOCAL_BALANCE_TREE_DEPTH: usize = 192;
 
 // TODO: This is basically the same as the nullifier tree, consider refactoring
 /// A commitment to the set of per-network nullifier trees maintained by the
@@ -34,27 +27,6 @@ where
     empty_hash_at_height: [H::Digest; LOCAL_BALANCE_TREE_DEPTH],
 }
 
-pub type LocalBalancePath<H> = SmtMerkleProof<H, LOCAL_BALANCE_TREE_DEPTH>;
-
-impl ToBits<192> for TokenInfo {
-    fn to_bits(&self) -> [bool; 192] {
-        let address_bytes = self.origin_token_address.0;
-        // Security: We assume here that `address_bytes` is a fixed-size array of
-        // 20 bytes. The following code could panic otherwise.
-        std::array::from_fn(|i| {
-            if i < 32 {
-                (*self.origin_network >> i) & 1 == 1
-            } else {
-                ((address_bytes[(i - 32) / 8]) >> (i % 8)) & 1 == 1
-            }
-        })
-    }
-}
-
-pub trait FromU256 {
-    fn from_u256(u: U256) -> Self;
-}
-
 impl<H> Default for LocalBalanceTree<H>
 where
     H: Hasher,
@@ -64,6 +36,7 @@ where
         Self::new()
     }
 }
+
 impl<H> LocalBalanceTree<H>
 where
     H: Hasher,
@@ -92,24 +65,12 @@ where
         res.root = root;
         res
     }
+}
 
-    // TODO: Consider batching the updates per network for efficiency
-    pub fn verify_and_update(
-        &mut self,
-        key: TokenInfo,
-        path_to_update: &LocalBalancePath<H>,
-        old_balance: U256,
-        new_balance: U256,
-    ) -> Result<(), ProofError> {
-        self.root = path_to_update
-            .verify_and_update(
-                key,
-                H::Digest::from_u256(old_balance),
-                H::Digest::from_u256(new_balance),
-                self.root,
-            )
-            .ok_or(ProofError::InvalidBalancePath)?;
-
-        Ok(())
+impl From<LocalBalanceTree<Keccak256Hasher>>
+    for pessimistic_proof_core::local_balance_tree::LocalBalanceTree<Keccak256Hasher>
+{
+    fn from(tree: LocalBalanceTree<Keccak256Hasher>) -> Self {
+        Self { root: tree.root }
     }
 }
