+++ crates/pessimistic-proof/src/local_exit_tree/tests.rs
@@ -1,9 +1,8 @@
-use hasher::Keccak256Hasher;
+use pessimistic_proof_core::{keccak::digest::Digest, local_exit_tree::hasher::Keccak256Hasher};
 use rs_merkle::{Hasher as MerkleHasher, MerkleTree};
 use tiny_keccak::{Hasher as _, Keccak};
 
 use super::*;
-use crate::keccak::digest::Digest;
 
 #[test]
 fn test_local_exit_tree_basic() {
